+++ src/config/commerce-picqer.php
@@ -13,5 +13,5 @@ return [
     //'orderStatusMapping' => [
     //    ['craft' => 1, 'picqer' => 'completed', 'changeTo' => 2],
     //],
-    //'fastStockUpdate' => false,
+    //'inventoryLocationId' => null,
 ];